control tcp

+++ tests/control_tcp_receiver.c
@@ -34,8 +34,8 @@ unsigned int total_msg = 500;
 unsigned int started_flag = 0;
 unsigned int stopped_flag = 0;
 
-unsigned int time_start = 0;
-unsigned int time_total = 0;
+double time_start = 0;
+double time_total = 0;
 unsigned int count_msg = 0;
 
 void* api_on_message(void* data)
